+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
  * Handle tabs `\t` with configurable tab size.
  * Added `TabSize` struct and related style builder method `tab_size`.
  * Added partial support for [ANSI escape codes](https://en.wikipedia.org/wiki/ANSI_escape_code).
+ * `Scrolling` vertical alignment
 
 ## Changed: